+++ CalculadoraImpostos_SergioDias.Services/ITaxCalculator.cs
@@ -4,7 +4,7 @@ namespace CalculadoraImpostos_SergioDias.Services
 {
     public interface ITaxCalculator
     {
-        public decimal TaxCalculation(decimal totalRevenue);
+        public double TaxCalculation(double totalRevenue);
         public bool RegisterTaxValue(Person person);
         public Person SearchTaxInfo(string cpf);
         public List<Person> ListTaxInfo();
